+++ fiddles/angular2-cli/fiddle-0036-EntAgGridToolTips/src/app/app.module.ts
@@ -1,0 +1,22 @@
+import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+import { HttpModule } from '@angular/http';
+import { AgGridModule } from "@ag-grid-community/angular";
+import { AppComponent } from './app.component';
+import { ComponentsModule } from './components/components.module';
+import { TooltipComponent } from './components/tooltip/tooltip.component';
+
+@NgModule({
+  declarations: [
+    AppComponent
+  ],
+  imports: [
+    BrowserModule,
+    HttpModule,
+    AgGridModule.withComponents([TooltipComponent]),
+    ComponentsModule
+  ],
+  providers: [],
+  bootstrap: [AppComponent]
+})
+export class AppModule { }