+++ fiddles/angular2-cli/fiddle-0036-EntAgGridToolTips/src/app/components/components.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { TooltipComponent } from './tooltip/tooltip.component';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  declarations: [TooltipComponent]
+})
+export class ComponentsModule { }